+++ lib/tir/defun.ml
@@ -161,7 +161,7 @@ let builtin_names : StringSet.t =
       "native_int_arr_make"; "native_int_arr_get"; "native_int_arr_set";
       "native_int_arr_length"; "native_int_arr_from_list"; "native_int_arr_to_list";
       "native_int_arr_map"; "native_int_arr_map2"; "native_int_arr_to_float_arr";
-      "native_int_arr_fold"; "native_int_arr_sum";
+      "native_int_arr_fold"; "native_int_arr_sum"; "native_int_arr_sort";
       "native_int_arr_min"; "native_int_arr_max"; "native_int_arr_sumsq_dev";
       "native_float_arr_make"; "native_float_arr_get"; "native_float_arr_set";
       "native_float_arr_length"; "native_float_arr_from_list"; "native_float_arr_to_list";